pyproject.toml
[project]
name = "consumer-complaints-ml-genai-ntier"
version = "0.1.0"
description = "AI-powered Policy Assistant using RAG with FastAPI, OpenAI, and Qdrant Vector Database"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Viswanatha Swamy P K"}
]
keywords = [
    "rag",
    "retrieval-augmented-generation",
    "enterprise-policy",
    "openai",
    "fastapi",
    "qdrant",
    "vector-database",
    "semantic-search"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business",
]

dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.1",
    "pydantic==2.6.1",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.1",
    "openai==1.12.0",
    "qdrant-client==1.7.3",
    "PyPDF2==3.0.1",
    "python-docx==1.1.0",
    "python-multipart==0.0.9",
    "pyjwt==2.8.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "email-validator==2.3.0",
    "python-jose[cryptography]==3.3.0",
    "psycopg2-binary==2.9.10",
    "sqlalchemy==2.0.36",
    "pyyaml==6.0.2",
    "httpx==0.27.2",
]

[project.optional-dependencies]
dev = [
    "black==24.10.0",
    "isort==5.13.2",
    "flake8==7.1.1",
    "mypy==1.14.1",
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "pytest-asyncio==0.25.2",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | __pycache__
  | \.pytest_cache
  | \.coverage
  | build
  | dist
  | node_modules
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["venv/*", ".venv/*", "*.pyc", "node_modules/*"]

[tool.pytest.ini_options]
testpaths = ["src/backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/backend/app"]
omit = ["*/tests/*", "*/test_*.py"]
